extracted game.ui, separated both main functions
