+++ game/application/pom.xml
@@ -15,25 +15,13 @@
     <description>This module bundles all the stuff together in a zircon application.</description>
 
     <properties>
-        <main.class>de.gleex.pltcmd.game.MainKt</main.class>
+        <main.class>de.gleex.pltcmd.game.application.MainKt</main.class>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>de.gleex.pltcmd.game</groupId>
-            <artifactId>communication</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>de.gleex.pltcmd.game</groupId>
-            <artifactId>options</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>de.gleex.pltcmd.model</groupId>
-            <artifactId>mapgeneration</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>de.gleex.pltcmd.model</groupId>
-            <artifactId>world</artifactId>
+            <artifactId>ui</artifactId>
         </dependency>
     </dependencies>
