+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/MapGenerationProgressController.kt
