+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/fragments/RadioSignalFragment.kt
@@ -1,6 +1,6 @@
 package de.gleex.pltcmd.game.ui.fragments
 
-import de.gleex.pltcmd.game.entities.TileRepository
+import de.gleex.pltcmd.game.ui.entities.TileRepository
 import de.gleex.pltcmd.model.radio.*
 import de.gleex.pltcmd.options.GameOptions
 import org.hexworks.cobalt.databinding.api.binding.bindPlusWith
