+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/mapgeneration/PreviewGenerationListener.kt
