+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/renderers/CoordinateDrawParameters.kt
